+++ test/SpecCustomDelimiters.hs
@@ -2,6 +2,13 @@ module SpecCustomDelimiters where
 
 import Language.Haskell.TH.Quote
 import PyF
+import PyF.Internal.QQ
 
 myCustomFormatter :: QuasiQuoter
-myCustomFormatter = fmtWithDelimiters ('@', '!')
+myCustomFormatter =
+  mkFormatter
+    "fmt"
+    ( defaultConfig
+        { delimiters = ('@', '!')
+        }
+    )